requirements.txt
# Core dependencies for GMShoot SOTA Analysis
streamlit>=1.28.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=1.5.0
Pillow>=9.5.0
opencv-python>=4.7.0
requests>=2.28.0
inference-sdk>=0.19.0

# Development dependencies
pytest>=7.2.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0